+++ src/lib/const/index.ts
@@ -1,5 +1,29 @@
 export const VENUE = {
-    NAME: 'Bóveda Secreta',
-    ADDRESS: 'San Antonio 705, Santiago. Región Metropolitana',
-    MAP_URL: 'https://maps.app.goo.gl/W95NsXm12SRUNKfW9'
-  }
+  NAME: "Bóveda Secreta",
+  ADDRESS: "San Antonio 705, Santiago. Región Metropolitana",
+  MAP_URL: "https://maps.app.goo.gl/W95NsXm12SRUNKfW9",
+};
+
+export const infoLanding = {
+  hero: {
+    title: "Clandestino",
+    description:
+      "¡Sumérgete en la clandestinidad de 5 Luchas! Embárcate en un viaje liberador que romperá las cadenas del estrés y la monotonía a través de la electrizante pasión de la Lucha Libre.",
+    image:
+      "https://res.cloudinary.com/dtj5xnlou/image/upload/f_auto,q_auto,w_1500,h_1001/v1/5LC/u69g3iwicq5jtlvyonwp",
+    video:
+      "https://res.cloudinary.com/dtj5xnlou/video/upload/f_auto:video,q_auto/yt1mvqeddyok3dqszmpz",
+  },
+  newsletter: {
+    title: "¡Suscríbete a nuestro Newsletter!",
+    description:
+      "Semanalmente te acercamos las noticias más destacadas del proyecto, los eventos venideros, las carteleras, los descuentos y mucho más.",
+    btn: "YA PO'",
+  },
+  cta: {
+    title: "¡Se parte!",
+    description: "Y súmate a la comunidad",
+    image:
+      "https://res.cloudinary.com/dtj5xnlou/image/upload/f_auto,q_auto,w_1500,h_1001/v1/5LC/bufh9o713u3otfclhnqq",
+  },
+};