+++ src/routes/(app)/+page.svelte
@@ -1,26 +1,19 @@
 <script lang="ts">
   import { urlForImage } from "$lib/config/sanity";
   import type { PageData } from "./$types";
-  import Header from "$lib/components/Hero.svelte";
-  import { PortableText } from "@portabletext/svelte";
-  import { LocaleConfig } from "$lib/utils/index";
   import { page } from "$app/stores";
   import { onMount } from "svelte";
-  import ModalTickets from "$lib/components/events/BatchModalTickets.svelte";
   import InfiniteScroll from "$lib/components/InfiniteScroll.svelte";
   import CardEvent from "$lib/components/events/CardEvent.svelte";
-  import DisclaimerModal from "$lib/components/DisclaimerModal.svelte";
-  import SingleEvent from "$lib/components/SingleEvent.svelte";
   import SuscribeForm from "$lib/components/SuscribeForm.svelte";
   import { TinySlider } from "svelte-tiny-slider";
-  import { ChevronRight } from "lucide-svelte";
-  import CardStaff from "$lib/components/CardStaff.svelte";
   import Slider from "$lib/components/equipo/Slider.svelte";
   import Faq from "$lib/components/Faq.svelte";
   import Hero from "$lib/components/Hero.svelte";
   import WhoWeAre from "$lib/components/WhoWeAre.svelte";
   import Statistics from "$lib/components/Statistics.svelte";
   import Video from "$lib/components/Video.svelte";
+  import { infoLanding } from "$lib/const";
 
   export let data: PageData;
 
@@ -60,32 +53,6 @@
 
     return () => window.removeEventListener("resize", checkScreenWidth);
   });
-
-  const info = {
-    SIEMPRE_BUENA_ONDA:
-      "https://res.cloudinary.com/dtj5xnlou/image/upload/f_auto,q_auto/v1/5LC/qnaz8aihcarkfaqjtipr",
-    INTENSIDAD:
-      "https://res.cloudinary.com/dtj5xnlou/image/upload/f_auto,q_auto/v1/5LC/mbyajlax7ea1hkuuwkbe",
-    BOVEDA_SECRETA:
-      "https://res.cloudinary.com/dtj5xnlou/image/upload/f_auto,q_auto/v1/5LC/bufh9o713u3otfclhnqq",
-  };
-
-  const infoHero = {
-    title: "Clandestino",
-    description:
-      "¡Ven a vivir la experiencia de 5 Luchas Clandestino! Acompañanos en la búsqueda de la liberación del estrés y la rutina diaria a través de la Lucha Libre. ",
-    image:
-      "https://res.cloudinary.com/dtj5xnlou/image/upload/f_auto,q_auto,w_1500,h_1001/v1/5LC/u69g3iwicq5jtlvyonwp",
-    video:
-      "https://res.cloudinary.com/dtj5xnlou/video/upload/f_auto:video,q_auto/yt1mvqeddyok3dqszmpz",
-  };
-
-  const infoCTA = {
-    title: "¡Se parte!",
-    description: "Y sumate a la comunidad",
-    image:
-      "https://res.cloudinary.com/dtj5xnlou/image/upload/f_auto,q_auto,w_1500,h_1001/v1/5LC/bufh9o713u3otfclhnqq",
-  };
 </script>
 
 <svelte:head>
@@ -127,7 +94,7 @@
     <InfiniteScroll />
   {/if}
   <Hero
-    info={infoHero}
+    info={infoLanding.hero}
     tag="h1"
     titleClass="text-6xl sm:text-8xl md:text-9xl"
   />
@@ -138,7 +105,7 @@
     <h2
       class="text-4xl font-bold text-white mask font-ibm italic mb-4 pl-4 xl:pl-0"
     >
-      Próximos Eventos <span
+      Próximos Eventos <span aria-hidden="true"
         class="font-[AtomicMarkerExtras] text-primary text-5xl">P</span
       >
     </h2>
@@ -148,13 +115,13 @@
           <CardEvent {event} />
         {/each}
       </TinySlider>
-      <SuscribeForm />
+      <SuscribeForm info={infoLanding.newsletter}/>
     {:else}
       <div class="flex items-start lg:gap-4 pl-4 xl:pl-0">
         {#each events as event}
           <CardEvent {event} />
         {/each}
-        <SuscribeForm />
+        <SuscribeForm info={infoLanding.newsletter} />
       </div>
     {/if}
   </main>
@@ -165,7 +132,7 @@
   <WhoWeAre />
   <Faq />
   <Hero
-    info={infoCTA}
+    info={infoLanding.cta}
     containerClass={"min-h-[450px] lg:min-h-[600px] container"}
     titleClass={"text-7xl"}
     descriptionClass={"font-ibm italic text-4xl"}